requirements.txt
# ==== API ====
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ==== YOLO ====
ultralytics==8.3.0
torch>=2.2.0
torchvision>=0.17.0

# ==== Image Processing ====
opencv-python==4.10.0.84
numpy>=1.26.0
Pillow>=10.3.0

# ==== OCR ====
paddlepaddle==2.6.0
paddleocr==2.7.0.3

# ==== Utils ====
pydantic==2.9.2
requests>=2.32.0
